change input type implemented, added component class, minor ui fix and improvements

+++ scripts/components/Note.js
@@ -1,35 +1,35 @@
+import Component from "../classes/Component.js";
 import { deleteCurrentNode } from "../db.js";
 import { ActionTypes } from "../enums.js";
 import { dispatch, initialState } from "../state.js";
-import {
-  autoResizeTextarea,
-  createEl,
-  select,
-  updateSaveButton,
-} from "../utils.js";
-import BlockArea from "./BlockArea.js";
+import { select } from "../utils.js";
 
 const mdCharacters = ["#", "*", "_"];
 
 export default function Note({ id, body }) {
-  let note = createEl("div");
-  note.classList.add("note");
+  let note = new Component({ el: "div", classList: ["note"] });
 
   let noteTitle = Object.values(body)[0].value.split("\n")[0];
 
-  let title = createEl("span");
-  title.classList.add("title");
-  title.innerText = `${
-    mdCharacters.includes(noteTitle[0])
-      ? noteTitle.slice(1, noteTitle.length).trim()
-      : noteTitle
-  }`;
+  let title = new Component({
+    el: "span",
+    classList: ["title"],
+    text: `${
+      mdCharacters.includes(noteTitle[0])
+        ? noteTitle.slice(1, noteTitle.length).trim()
+        : noteTitle
+    }`,
+  });
 
-  let dltBtn = createEl("i");
-  dltBtn.classList.add("si-x", "btn_icon");
-  dltBtn.style.fontSize = "20px";
+  let dltBtn = new Component({
+    el: "i",
+    classList: ["si-x", "btn_icon"],
+    styles: { fontSize: "20px" },
+  });
 
-  title.addEventListener("click", () => {
+  title.addListener(Component.ListenerTypes.Click, () => {
+    initialState.tabs.toggle("#editor");
+
     dispatch({
       type: ActionTypes.SetEditorData,
       payload: { editorData: body },
@@ -49,55 +49,47 @@ export default function Note({ id, body }) {
       type: ActionTypes.UpdateCurrentNoteSavedState,
       payload: { currentNoteSaved: true },
     });
-
-    select(".inputs").innerHTML = "";
-
-    for (const key in body) {
-      select(".inputs").append(
-        BlockArea({
-          id: key,
-          type: body[key].type,
-          defaultValue: body[key].value,
-        })
-      );
-    }
-
-    initialState.tabs.toggle("#editor");
-
-    updateSaveButton();
-
-    autoResizeTextarea();
   });
 
-  dltBtn.addEventListener("click", async () => {
-    await deleteCurrentNode(id);
+  dltBtn.addListener(Component.ListenerTypes.Click, async () => {
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#d33",
+      cancelButtonColor: "#3085d6",
+      confirmButtonText: "Yes, delete it!",
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        await deleteCurrentNode(id);
 
-    dispatch({
-      type: ActionTypes.SetEditorData,
-      payload: { editorData: {} },
+        dispatch({
+          type: ActionTypes.SetEditorData,
+          payload: { editorData: {} },
+        });
+
+        dispatch({
+          type: ActionTypes.SetCurrentNoteId,
+          payload: { currentNoteId: null },
+        });
+
+        dispatch({
+          type: ActionTypes.SetPreviouslyCreatedNoteOpened,
+          payload: { previouslyCreatedNoteOpened: false },
+        });
+
+        dispatch({
+          type: ActionTypes.UpdateCurrentNoteSavedState,
+          payload: { currentNoteSaved: false },
+        });
+
+        select(".content").innerHTML = "";
+      }
     });
-
-    dispatch({
-      type: ActionTypes.SetCurrentNoteId,
-      payload: { currentNoteId: null },
-    });
-
-    dispatch({
-      type: ActionTypes.SetPreviouslyCreatedNoteOpened,
-      payload: { previouslyCreatedNoteOpened: false },
-    });
-
-    dispatch({
-      type: ActionTypes.UpdateCurrentNoteSavedState,
-      payload: { currentNoteSaved: false },
-    });
-
-    select(".inputs").innerHTML = "";
-    select(".content").innerHTML = "";
   });
 
-  note.append(title);
-  note.append(dltBtn);
+  note.addChildren([title, dltBtn]);
 
-  return note;
+  return note.getDomNode();
 }
